chore(e2e): retire legacy pre-baseline E2E tree; make baseline self-sufficient [INT-966]
Delete everything under tests/e2e/ outside baseline/ (adapters/, scenarios/, fixtures/, helpers/, root conftest.py, settings.py, _constants.py); keep only tests/e2e/__init__.py and the baseline toolkit + suite.

The remaining legacy scenarios are dropped, not migrated: their unique behavior is already covered by unit tests (agno history guard in test_history_guard.py; reasoning->thought translation in test_adapter.py) or is deliberately out of baseline scope (mid-tool-call re-execution, per test_rehydration_idempotency.py's docstring); participant removal is covered by test_participant_management.py. The one genuinely-uncovered behavior (implicit subject-scope memory inference) was migrated separately.

Migrate the two load-bearing side effects the baseline suite silently inherited from the legacy root conftest/settings, so it now stands alone: the E2E_TIMEOUT collection hook (including @pytest.mark.timeout(extra=N) resolution) into baseline/conftest.py, and load_dotenv(.env.test) into baseline/settings.py so framework LLM SDKs read provider keys from os.environ. Point CLAUDE.md's Testing Structure at baseline as the sole E2E suite.
