+++ tests/e2e/baseline/settings.py
@@ -9,8 +9,19 @@ required to run. Add a new subclass + nested field as new concerns appear
 
 from __future__ import annotations
 
+from pathlib import Path
+
+from dotenv import load_dotenv
 from pydantic import AliasChoices, Field, model_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
+
+# Load .env.test into os.environ (idempotent, non-overriding) so the framework
+# LLM SDKs (anthropic, openai, google-genai) that read their provider keys
+# straight from the environment pick them up. pydantic-settings' ``env_file``
+# only populates this module's own fields, not ``os.environ``, so the SDKs — and
+# any adapter/model the toolkit builds — need this explicit load. Imported before
+# any fixture constructs a client or adapter, so keys are present in time.
+load_dotenv(Path(__file__).parents[3] / ".env.test", override=False)
 
 
 class BandEndpoints(BaseSettings):
