--- tests/e2e/fixtures/__init__.py
@@ -1,7 +1,0 @@
-"""E2E fixture modules, imported into ``tests/e2e/conftest.py``'s namespace.
-
-Split by concern: ``clients`` (config + REST/WS clients), ``rooms`` (room
-allocation + agent identity), ``memory`` (memory-test toolkit).
-"""
-
-from __future__ import annotations
